pull master

+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/metrics/DoubleAvgMetrics.java
@@ -34,11 +34,11 @@ public abstract class DoubleAvgMetrics extends Metrics implements DoubleValueHol
     protected static final String VALUE = "value";
 
     @Getter @Setter @Column(columnName = SUMMATION) private double summation;
-    @Getter @Setter @Column(columnName = COUNT) private int count;
+    @Getter @Setter @Column(columnName = COUNT) private long count;
     @Getter @Setter @Column(columnName = VALUE, isValue = true, function = Function.Avg) private double value;
 
     @Entrance
-    public final void combine(@SourceFrom double summation, @ConstOne int count) {
+    public final void combine(@SourceFrom double summation, @ConstOne long count) {
         this.summation += summation;
         this.count += count;
     }
